devfreq: Allow bw_hwmon resume with zero resume freq
Currently bw_hwmon governor does not allow the device
to be resumed if the resume_freq is set to zero. But there
can be devices that vote for zero frequency when there is
no bandwidth on the device. This does not mean that the device
should not be resume from suspend. Remove the check with prevents
resume, instead add a check in the devfreq framework to check if
device is suspended or not before calling the governor resume.

Change-Id: Ib82a6a36308aee52e8bb989fddca92265de6c4a4
Signed-off-by: Rama Aparna Mallavarapu <aparnam@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>
(cherry picked from commit 53cfc53ab07b33bf0724e70134e75ce00c23cdc3)