msm: kgsl: Change error log to info log in gpumem_alloc_entry()
The largest allowable alignment for a GPU object is 32MB in driver.
If the alignment is more than supported, we restrict the alignment.
So what we dump in kernel log related to alignment restriction is
info to user not an error.

Change-Id: I74d1f036c2669c67cd3fb0e6e1b1cc4c67aa2815
Signed-off-by: Hareesh Gundu <hareeshg@codeaurora.org>
Signed-off-by: Sravankumar bijili <csbijil@codeaurora.org>
(cherry picked from commit 986e4af6577beb73908da26807f3b98a6ab4bbaf)