usb: dwc3: Fix enumeration failure with reset during transfers
If the host initiates a BUS reset when SETUP packet
is in progress, it is possible that the trb_enqueue
is set to 1 and ep0_bounced is set to true in handling
the SETUP packet which was interrupted with the BUS reset.

With the trb_enqueue set to 1, instead of the first TRB
the second TRB is queued but the START_TRANSFER is issued
on the first TRB which does not have HWO is set leading
to the enumeration failure, since there is no available
TRB to handle the SETUP packet from the host.

Fix this by resetting the trb_enqueue and ep0_bounced
to prevent the enumeration failure.

Change-Id: I73a4ca0a970124714acc511ab507d1411d197709
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
(cherry picked from commit b8a4d2637fb392e8a6a22a0b25be685934e12a8b)