msm: camera: cci: Fix incorrect use of cci config ioctl
The cci configuration will be transitioned to a new API that does not
require routing through the v4l layer. This is work-in-progrss so in
the mean time prevent the device from being exposed as configurable
from userspace.
The ioctl will still be exposed to kernel users so fix the arg size
as well. We want size of struct not pointer.

Change-Id: I1eeb58b8355a81fc36b9845e2096060b03f1debe
Signed-off-by: Om Parkash <oparkash@codeaurora.org>
(cherry picked from commit 4da13078e628f76c5e3a3dee9f0fb3d01e51de34)